requirements.txt
pandas>=1.5.0
google-generativeai>=0.3.0
requests>=2.28.0
python-dotenv>=0.19.0
numpy>=1.21.0
beautifulsoup4>=4.11.0
supabase>=2.0.0
